Analytics classの作成
