+++ app/build.gradle
@@ -1,12 +1,11 @@
 plugins {
+    id "com.android.application"
     id "org.jetbrains.kotlin.android"
     id "kotlin-kapt"
     id "dagger.hilt.android.plugin"
     id "com.google.gms.google-services"
     id 'com.google.firebase.crashlytics'
 }
-
-apply plugin: "com.android.application"
 
 def getGoogleFormId() {
     def propFile = project.rootProject.file('local.properties')
