+++ app/src/main/java/io/github/rikuyu/contactlensreminder/data/local/alarm/tickdown/TickDownAlarmManager.kt
@@ -4,19 +4,18 @@ import android.app.AlarmManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import com.google.firebase.analytics.ktx.analytics
-import com.google.firebase.analytics.ktx.logEvent
-import com.google.firebase.ktx.Firebase
 import io.github.rikuyu.contactlensreminder.data.local.sharedpreferences.SharedPreferencesManager
 import io.github.rikuyu.contactlensreminder.data.util.ChangeAppIconService
+import io.github.rikuyu.contactlensreminder.data.util.FirebaseLogEvent
 import java.text.SimpleDateFormat
 import java.util.*
 import javax.inject.Inject
 
-class TickDownAlarmManager @Inject constructor(private val context: Context) {
-
+class TickDownAlarmManager @Inject constructor(
+    private val context: Context,
+    private val firebaseLogEvent: FirebaseLogEvent
+) {
     private val changeAppIconService: ChangeAppIconService = ChangeAppIconService(context)
-    private val sharedPreferencesManager = SharedPreferencesManager(context)
     private val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
 
     fun initAlarm() {
@@ -41,7 +40,7 @@ class TickDownAlarmManager @Inject constructor(private val context: Context) {
             calendar.timeInMillis,
             pendingIntent
         )
-        logInitTickDownEvent()
+        firebaseLogEvent.logInitTickDownEvent()
     }
 
     fun cancelAlarm() {
@@ -54,32 +53,10 @@ class TickDownAlarmManager @Inject constructor(private val context: Context) {
         )
         alarmManager.cancel(pendingIntent)
         changeAppIconService.changeAppIcon(false, null)
-        logCancelTickDownEvent()
-    }
-
-    private fun logInitTickDownEvent() {
-        val uuid = sharedPreferencesManager.getUuid() ?: return
-        val lensPeriod = sharedPreferencesManager.getContactLensPeriod().toString()
-        val remainingDay = sharedPreferencesManager.getContactLensRemainingDays().toString()
-
-        val firebaseAnalytics = Firebase.analytics
-        firebaseAnalytics.logEvent("init_tick_down_alarm_event") {
-            param(UUID, uuid)
-            param(LENS_PERIOD, lensPeriod)
-            param(REMAINING_DAY, remainingDay)
-        }
-    }
-
-    private fun logCancelTickDownEvent() {
-        val uuid = sharedPreferencesManager.getUuid() ?: return
-        val firebaseAnalytics = Firebase.analytics
-        firebaseAnalytics.logEvent("cancel_tick_down_alarm_event") { param(UUID, uuid) }
+        firebaseLogEvent.logEvent("cancel_tick_down_alarm_event")
     }
 
     companion object {
         private const val REQUEST_CODE = 7777
-        private const val LENS_PERIOD = "lens_period"
-        private const val REMAINING_DAY = "remaining_day"
-        private const val UUID = "uuid"
     }
 }
